added creator of the post

+++ backend/routes/posts.js
@@ -36,7 +36,8 @@ router.post("",checkAuth,multer({ storage: storage }).single("image"),(req, res,
     const post = new Post({
       title: req.body.title,
       content: req.body.content,
-      imagePath: url + "/images/" + req.file.filename
+      imagePath: url + "/images/" + req.file.filename,
+      creator: req.userData.userId
     });
     post.save().then(createdPost => {
       res.status(201).json({
@@ -60,11 +61,16 @@ router.put("/:id",checkAuth,multer({ storage: storage }).single("image"),(req, r
       _id: req.body.id,
       title: req.body.title,
       content: req.body.content,
-      imagePath: imagePath
+      imagePath: imagePath,
+      creator: req.userData.userId
     });
-    console.log(post);
-    Post.updateOne({ _id: req.params.id }, post).then(result => {
-      res.status(200).json({ message: "Update successful!" });
+    Post.updateOne({ _id: req.params.id,creator: req.userData.userId }, post).then(result => {
+      if (result.modifiedCount > 0) {
+        res.status(200).json({ message: "Update successful!" });
+      } else {
+        res.status(401).json({ message: "Not authorized!" });
+      }
+      
     });
   }
 );
@@ -102,9 +108,12 @@ router.get("/:id", (req, res, next) => {
 });
 
 router.delete("/:id", checkAuth,(req, res, next) => {
-  Post.deleteOne({ _id: req.params.id }).then(result => {
-    console.log(result);
-    res.status(200).json({ message: "Post deleted!" });
+  Post.deleteOne({ _id: req.params.id,creator: req.userData.userId  }).then(result => {
+    if (result.deletedCount > 0) {
+      res.status(200).json({ message: "Deletion Succssful" });
+    }else {
+      res.status(401).json({ message: "Not authorized!" });
+    }
   });
 });
